requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
certifi==2019.11.28
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
imageio==2.8.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
natsort==7.0.1
nbconvert==5.6.1
nbformat==5.0.5
networkx==2.4
nibabel==3.0.1
nilearn==0.6.2
notebook==6.0.3
numpy==1.18.1
opencv-python==4.2.0.32
openpyxl==3.0.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
progress==1.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
SimpleITK==1.2.4
six==1.14.0
sklearn==0.0
terminado==0.8.3
testpath==0.4.4
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
zipp==3.1.0
